feat!: migrate to Nuxt 4 and NuxtUI v4 with new directory structure
BREAKING CHANGE: Upgraded from Nuxt 3 to Nuxt 4 and NuxtUI v3 to v4

- Upgraded nuxt from 3.15.0 to 4.1.2
- Upgraded @nuxt/ui from v3.0.0-alpha.10 to v4.0.1
- Migrated to Nuxt 4 directory structure with app/ folder
- Moved all source code (components, composables, pages, stores, types) to app/
- Updated nuxt.config.ts with future.compatibilityVersion: 4
- Updated vitest.config.ts paths for new structure
- Refactored modal management to remove deprecated useModal API
- Implemented state-based modal handling in composables and components
- Updated all configuration files (eslint, vitest) for new paths
- Updated documentation (README.md, development.md) with Nuxt 4 references
- Added comprehensive MIGRATION_SUMMARY.md documentation

All functionality verified working: dev server, production build, and tests passing.
