+++ app/components/app/chat/ModelDropdown.vue
