+++ app/components/app/models/ModelCard.vue
