+++ app/components/app/models/actions/ConfirmCopy.vue
