+++ app/components/app/models/actions/ConfirmDelete.vue
