+++ app/components/app/models/details/ModelFile.vue
