+++ app/components/app/settings/actions/ConfirmReset.vue
