+++ app/components/app/settings/connection/AirplaneModeToggle.vue
