+++ app/components/app/settings/connection/OllamaURL.vue
