+++ app/components/base/BaseConfirmationModal.vue
