+++ app/composables/useOllama.ts
