+++ app/pages/index.vue
