+++ app/stores/chat.ts
