+++ app/stores/settings.ts
