+++ app/types/ollama.ts
