+++ nuxt.config.ts
@@ -1,12 +1,23 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
-	modules: ['@nuxt/ui', '@pinia/nuxt', '@vueuse/nuxt', '@nuxt/eslint'],
-	css: ['~/assets/main.css'],
-	compatibilityDate: '2024-11-01',
-	devtools: { enabled: true },
-	ui: {
-		theme: {
-			colors: ['primary', 'secondary', 'success', 'info', 'warning', 'error', 'neutral'],
-		},
-	},
+  modules: ["@nuxt/ui", "@pinia/nuxt", "@vueuse/nuxt", "@nuxt/eslint"],
+  css: ["~/assets/main.css"],
+  compatibilityDate: "2024-11-01",
+  devtools: { enabled: true },
+  future: {
+    compatibilityVersion: 4,
+  },
+  ui: {
+    theme: {
+      colors: [
+        "primary",
+        "secondary",
+        "success",
+        "info",
+        "warning",
+        "error",
+        "neutral",
+      ],
+    },
+  },
 });
