+++ package.json
@@ -13,11 +13,12 @@
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@nuxt/ui": "^3.0.0-alpha.10",
+    "@nuxt/ui": "^4.0.0",
     "@pinia/nuxt": "^0.9.0",
     "@types/uuid": "^10.0.0",
     "@vueuse/nuxt": "^12.3.0",
-    "nuxt": "^3.15.0",
+    "caniuse-lite": "^1.0.30001748",
+    "nuxt": "^4.1.2",
     "ollama": "^0.5.11",
     "pinia": "^2.3.0",
     "uuid": "^11.0.5",
